+++ src/style/themes/base/base-theme.config.js
@@ -6,6 +6,8 @@ export default (palette) => {
   return {
     name: 'base',
 
+    spacing: 8,
+
     colors: {
       // main
       base: palette.productGreen,
@@ -39,6 +41,11 @@ export default (palette) => {
       destructive: {
         hover: palette.errorRedShade(20)
       }
+    },
+
+    accordion: {
+      border: palette.slateTint(80),
+      background: palette.slateTint(90)
     },
 
     menu: {